# swagger.yaml
swagger: '2.0'
info:
  description: 'HuyIT Bloodline DNA Testing Service Management System API docs'
  version: '1.0.0'
  title: 'Bloodline DNA Testing Service Management System'
  termOfService: 'HuyIT'
  contact:
    email: 'huyit2003@gmail.com'
  license:
    name: 'MIT'
    url: 'https://opensource.org/licenses/MIT'
basePath: '/api'
...
schemes:
- 'https'
- 'http'

# Declare security scheme
securityDefinitions:
  Bearer:
    type: 'apiKey'
    name: 'Authorization'
    in: 'header'
    description: 'Bearer token for authorization'
    x-tokenName: 'token'
    x-tokenPrefix: 'Bearer '
paths:
...
  /users/create:
    post:
      tags:
      - 'users'
      summary: 'Create user (Admin only)'
      description: 'Create a new user (Admin only)'
      operationId: "createUser"
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'body'
        name: 'body'
        description: 'User data'
        required: true
        schema: 
          $ref: '#/definitions/RegisterDto'
      responses:
        '201':
          description: 'User created successfully'
          schema:
            $ref: '#/definitions/UserResponse'
        '400':
          description: 'Invalid input'
        '401':
          description: 'Unauthorized'
...
  /users/{id}:
    get:
      tags:
      - 'users'
      summary: 'Get user by ID'
      description: 'Get user information by ID'
      operationId: "getUserById"
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'path'
        name: 'id'
        description: 'User ID'
        required: true
        type: 'string'
      responses:
        '200':
          description: 'User information'
          schema:
            $ref: '#/definitions/UserResponse'
        '400':
          description: 'Invalid ID'
        '404':
          description: 'User not found'
    put:
      tags:
      - 'users'
      summary: 'Update user'
      description: 'Update user information'
      operationId: "updateUser"
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'path'
        name: 'id'
        description: 'User ID'
        required: true
        type: 'string'
      - in: 'body'
        name: 'body'
        description: 'Updated user data'
        required: true
        schema: 
          $ref: '#/definitions/UpdateUserDto'
      responses:
        '200':
          description: 'User updated successfully'
          schema:
            $ref: '#/definitions/UserResponse'
        '400':
          description: 'Invalid input'
        '401':
          description: 'Unauthorized'
    delete:
      tags:
      - 'users'
      summary: 'Delete user (Admin only)'
      description: 'Delete user by ID (soft delete)'
      operationId: "deleteUser"
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'path'
        name: 'id'
        description: 'User ID'
        required: true
        type: 'string'
      responses:
        '200':
          description: 'User deleted successfully'
        '400':
          description: 'Invalid ID'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User not found'
  /users/change-password:
    put:
      tags:
      - 'users'
      summary: 'Change password'
      description: 'Change user password'
      operationId: "changePassword"
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'body'
        name: 'body'
        description: 'Password change data'
        required: true
        schema: 
          $ref: '#/definitions/ChangePasswordDto'
      responses:
        '200':
          description: 'Password changed successfully'
        '400':
          description: 'Invalid input'
        '401':
          description: 'Unauthorized'
...
  /service/create:
    post:
      tags:
      - 'services'
      summary: 'Create new service (Admin, Manager only)'
      description: 'Create a new service '
      operationId: "createService"
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'body'
        name: 'body'
        description: 'Service information'
        required: true
        schema: 
          $ref: '#/definitions/CreateServiceDto'
      responses:
        '201':
          description: 'Service created successfully'
          schema:
            $ref: '#/definitions/ServiceResponse'
        '400':
          description: 'Invalid input'
        '401':
          description: 'Unauthorized'
  /service/search:
    get:
      tags:
      - 'services'
      summary: 'Search and filter services (All roles)'
      description: 'Search services with pagination and filtering'
      operationId: "getServices"
      produces:
      - 'application/json'
      parameters:
      - in: 'query'
        name: 'type'
        description: 'Filter by service type (civil, administrative)'
        required: false
        type: 'string'
      - in: 'query'
        name: 'sample_method'
        description: 'Filter by sample method (self_collected, facility_collected, home_collected)'
        required: false
        type: 'string'
      - in: 'query'
        name: 'is_active'
        description: 'Filter by active status'
        required: false
        type: 'boolean'
      - in: 'query'
        name: 'min_price'
        description: 'Minimum price'
        required: false
        type: 'number'
      - in: 'query'
        name: 'max_price'
        description: 'Maximum price'
        required: false
        type: 'number'
      - in: 'query'
        name: 'pageNum'
        description: 'Page number'
        required: false
        type: 'integer'
        default: 1
      - in: 'query'
        name: 'pageSize'
        description: 'Items per page'
        required: false
        type: 'integer'
        default: 10
      - in: 'query'
        name: 'keyword'
        description: 'Search by name or description'
        required: false
        type: 'string'
      - in: 'query'
        name: 'sort_by'
        description: 'Sort field (name, price, created_at, estimated_time)'
        required: false
        type: 'string'
        default: 'created_at'
      - in: 'query'
        name: 'sort_order'
        description: 'Sort order (asc, desc)'
        required: false
        type: 'string'
        default: 'desc'
      - in: 'query'
        name: 'start_date'
        description: 'Filter by creation date (start)'
        required: false
        type: 'string'
        format: 'date'
      - in: 'query'
        name: 'end_date'
        description: 'Filter by creation date (end)'
        required: false
        type: 'string'
        format: 'date'
      responses:
        '200':
          description: 'Service list with pagination'
          schema:
            $ref: '#/definitions/ServicePaginationResponse'
        '401':
          description: 'Unauthorized'
  /service/{id}:
    get:
      tags:
      - 'services'
      summary: 'Get service by ID (All roles)'
      description: 'Retrieve service details by ID'
      operationId: "getServiceById"
      produces:
      - 'application/json'
      parameters:
      - in: 'path'
        name: 'id'
        description: 'Service ID'
        required: true
        type: 'string'
      responses:
        '200':
          description: 'Service details'
          schema:
            $ref: '#/definitions/ServiceResponse'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Service not found'
          
    put:
      tags:
      - 'services'
      summary: 'Update service (Admin, Manager only)'
      description: 'Update service information'
      operationId: "updateService"
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'path'
        name: 'id'
        description: 'Service ID'
        required: true
        type: 'string'
      - in: 'body'
        name: 'body'
        description: 'Updated service information'
        required: true
        schema: 
          $ref: '#/definitions/UpdateServiceDto'
      responses:
        '200':
          description: 'Service updated successfully'
          schema:
            $ref: '#/definitions/ServiceResponse'
        '400':
          description: 'Invalid input'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Service not found'

    delete:
      tags:
      - 'services'
      summary: 'Delete service (Admin, Manager only)'
      description: 'Delete service (soft delete)'
      operationId: "deleteService"
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'path'
        name: 'id'
        description: 'Service ID'
        required: true
        type: 'string'
      responses:
        '200':
          description: 'Service deleted successfully'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Service not found'
  /service/appointments:
    get:
      tags:
      - 'services'
      summary: 'Get services by appointment criteria (Admin, Manager, Staff only)'
      description: 'Filter services based on related appointments'
      operationId: "getServicesByAppointment"
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'query'
        name: 'status'
        description: 'Filter by appointment status'
        required: false
        type: 'string'
      - in: 'query'
        name: 'start_appointment_date'
        description: 'Filter by appointment date (start)'
        required: false
        type: 'string'
        format: 'date'
      - in: 'query'
        name: 'end_appointment_date'
        description: 'Filter by appointment date (end)'
        required: false
        type: 'string'
        format: 'date'
      - in: 'query'
        name: 'appointment_type'
        description: 'Filter by appointment type (self, facility, home)'
        required: false
        type: 'string'
      - in: 'query'
        name: 'customer_id'
        description: 'Filter by customer ID'
        required: false
        type: 'string'
      - in: 'query'
        name: 'staff_id'
        description: 'Filter by staff ID'
        required: false
        type: 'string'
      - in: 'query'
        name: 'collection_address'
        description: 'Filter by collection address'
        required: false
        type: 'string'
      - in: 'query'
        name: 'pageNum'
        description: 'Page number'
        required: false
        type: 'integer'
        default: 1
      - in: 'query'
        name: 'pageSize'
        description: 'Items per page'
        required: false
        type: 'integer'
        default: 10
      responses:
        '200':
          description: 'Service list filtered by appointment data'
          schema:
            $ref: '#/definitions/ServicePaginationResponse'
        '401':
          description: 'Unauthorized'
...
  /service/statistics:
    get:
      tags:
      - 'services'
      summary: 'Get service statistics (Admin, Manager)'
      description: 'Count services grouped by type, status, and sample method'
      operationId: "countServicesByType"
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'query'
        name: 'is_active'
        description: 'Filter by active status'
        required: false
        type: 'boolean'
      - in: 'query'
        name: 'keyword'
        description: 'Search by name or description'
        required: false
        type: 'string'
      responses:
        '200':
          description: 'Service statistics'
          schema:
            type: 'object'
            properties:
              total:
                type: 'integer'
                description: 'Total number of services'
              byType:
                type: 'object'
                description: 'Services counted by type'
                properties:
                  civil:
                    type: 'integer'
                  administrative:
                    type: 'integer'
              byStatus:
                type: 'object'
                description: 'Services counted by status'
                properties:
                  active:
                    type: 'integer'
                  inactive:
                    type: 'integer'
              bySampleMethod:
                type: 'object'
                description: 'Services counted by sample method'
                properties:
                  self_collected:
                    type: 'integer'
                  facility_collected:
                    type: 'integer'
                  home_collected:
                    type: 'integer'
        '401':
          description: 'Unauthorized'
  /service/{id}/status:
    patch:
      tags:
      - 'services'
      summary: 'Change service status (Admin, Manager)'
      description: 'Activate or deactivate a service'
      operationId: "changeServiceStatus"
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      security:
        - Bearer: []
      parameters:
      - in: 'path'
        name: 'id'
        description: 'Service ID'
        required: true
        type: 'string'
      - in: 'body'
        name: 'body'
        description: 'Status change data'
        required: true
        schema:
          type: 'object'
          properties:
            is_active:
              type: 'boolean'
              description: 'New active status'
          required:
            - is_active
      responses:
        '200':
          description: 'Service status changed successfully'
          schema:
            $ref: '#/definitions/ServiceResponse'
        '400':
          description: 'Invalid input or service is deleted'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Service not found'
definitions:
...
  UserResponse:
    type: object
    properties:
      _id:
        type: string
        description: "User ID"
      first_name:
        type: string
        description: "User's first name"
      last_name:
        type: string
        description: "User's last name"
      email:
        type: string
        format: email
        description: "User's email address"
      google_id:
        type: string
        description: "Google ID (if applicable)"
      role:
        type: string
        enum: ["admin", "manager", "staff", "customer"]
        description: "User role"
      status:
        type: boolean
        description: "User account status"
      phone_number:
        type: string
        description: "User's phone number"
      avatar_url:
        type: string
        description: "URL to user's avatar"
      dob:
        type: string
        format: date
        description: "Date of birth"
      is_verified:
        type: boolean
        description: "Email verification status"
      created_at:
        type: string
        format: date-time
        description: "Account creation date"
      updated_at:
        type: string
        format: date-time
        description: "Account last update date"
  RegisterDto:
    type: object
    properties:
      google_id:
        type: string
        description: "Google ID (optional)"
      first_name:
        type: string
        description: "User's first name"
      last_name:
        type: string
        description: "User's last name"
      email:
        type: string
        format: email
        description: "User's email address"
      password:
        type: string
        minLength: 6
        description: "User's password"
      role:
        type: string
        enum: ["customer", "manager", "staff", "admin"]
        description: "User role"
      phone_number:
        type: string
        description: "User's phone number"
      avatar_url:
        type: string
        description: "URL to user's avatar"
      dob:
        type: string
        format: date
        description: "Date of birth"
      address:
        type: string
        description: "User's address"
      gender:
        type: string
        description: "User's gender"
    required:
      - first_name
      - last_name
      - email
      - password
      - phone_number
      - dob
  UpdateUserDto:
    type: object
    properties:
      first_name:
        type: string
        description: "User's first name"
      last_name:
        type: string
        description: "User's last name"
      phone_number:
        type: string
        description: "User's phone number"
      avatar_url:
        type: string
        description: "URL to user's avatar"
      dob:
        type: string
        format: date
        description: "Date of birth"
      address:
        type: string
        description: "User's address"
      gender:
        type: string
        description: "User's gender"
    required:
      - first_name
      - last_name
  ChangePasswordDto:
    type: object
    properties:
      user_id:
        type: string
        description: "User ID"
      old_password:
        type: string
        description: "Current password"
        minLength: 6
      new_password:
        type: string
        description: "New password"
        minLength: 6
    required:
      - user_id
      - old_password
      - new_password
...
  CreateServiceDto:
    type: object
    properties:
      name:
        type: string
        description: "Service name"
        example: "Dân Sự"
      description:
        type: string
        description: "Service description"
        example: "Dịch vụ dân sự"
      price:
        type: number
        description: "Service price"
        example: 1500000
      type:
        type: string
        enum: ["civil", "administrative"]
        description: "Service type"
        example: "civil"
      sample_method:
        type: string
        enum: ["self_collected", "facility_collected", "home_collected"]
        description: "Sample collection method"
        example: "home_collected"
      estimated_time:
        type: number
        description: "Estimated time in hours"
        example: 48
    required:
      - name
      - description
      - price
      - type
      - sample_method
      - estimated_time
  UpdateServiceDto:
    type: object
    properties:
      name:
        type: string
        description: "Service name"
      description:
        type: string
        description: "Service description"
      price:
        type: number
        description: "Service price"
      type:
        type: string
        enum: ["civil", "administrative"]
        description: "Service type"
      sample_method:
        type: string
        enum: ["self_collected", "facility_collected", "home_collected"]
        description: "Sample collection method"
      estimated_time:
        type: number
        description: "Estimated time in hours"
    required:
      - name
      - description
      - price
      - type
      - sample_method
      - estimated_time
  ServiceResponse:
    type: object
    properties:
      _id:
        type: string
        description: "Service ID"
      name:
        type: string
        description: "Service name"
      description:
        type: string
        description: "Service description"
      price:
        type: number
        description: "Service price"
      type:
        type: string
        enum: ["civil", "administrative"]
        description: "Service type"
      sample_method:
        type: string
        enum: ["self_collected", "facility_collected", "home_collected"]
        description: "Sample collection method"
      estimated_time:
        type: number
        description: "Estimated time in hours"
      is_active:
        type: boolean
        description: "Service status"
      created_at:
        type: string
        format: date-time
        description: "Creation date"
      updated_at:
        type: string
        format: date-time
        description: "Last update date"
  ServicePaginationResponse:
    type: object
    properties:
      pageData:
        type: array
        items:
          $ref: '#/definitions/ServiceResponse'
      pageInfo:
        type: object
        properties:
          totalItems:
            type: integer
            description: "Total number of items"
          totalPages:
            type: integer
            description: "Total number of pages"
          pageNum:
            type: integer
            description: "Current page number"
          pageSize:
            type: integer
            description: "Number of items per page"
...